services/meteorologist/requirements.txt
anyio==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
python-dotenv==0.19.2
rfc3986==1.5.0
sniffio==1.2.0
tomli==2.0.1
Werkzeug==2.0.3
